Introduce protocols `IntervalMember` and `Countable`
Previously, `Interval.Member` must be `Hashable & Comparable`, which (was too restrictive and) caused everything defined in `Comparable`'s extensions to be statically dispatched for `Member`. `IntervalMember` defines the same things as its requirement, which allows dynamic dispatch for `Member`. See documentation for more details.

The previous `Strideable` extension was incorrect, and it's intent is reimplemented with `Countable`. Likewise, see documentation for more details.

Additionally, the requirement that `Interval` is `Hashable` is too restrictive. It's  lowered to `Equatable` instead.

- FIXME: Fix the test case for `Countable`'s default implementation.